+++ src/transformations/_manager.js
@@ -16,7 +16,7 @@ export default class TransformationManager {
   }
 
   //picks a random transformation out of a list of transformation names
-  pickOne(list) {
+  pickOne(list, potency = 0.5) {
     list = list
       .map(name => this._TRANSFORMATIONS.find(item => item.name === name))
       .filter(item => item)
@@ -34,7 +34,7 @@ export default class TransformationManager {
 
     const roll = Math.random()
 
-    if (roll < choice.chance) {
+    if (roll < potency) {
       return choice.apply()
     } else {
       return ""