+++ src/transformations/minotaur-cum.js
@@ -13,17 +13,20 @@ export default class MinotaurCum extends Transformation {
   }
 
   apply() {
-    return this.owner.transform.pickOne([
-      "grow four breasts",
-      "increase breast size",
-      "increase breast milk",
-      "increase penis size",
-      "change tail to bovine",
-      "grow bovine tail",
-      "grow a bovine udder",
-      "increase udder size",
-      "increase nipple size",
-      "grow an extra nipple"
-    ])
+    return this.owner.transform.pickOne(
+      [
+        "grow four breasts",
+        "increase breast size",
+        "increase breast milk",
+        "increase penis size",
+        "change tail to bovine",
+        "grow bovine tail",
+        "grow a bovine udder",
+        "increase udder size",
+        "increase nipple size",
+        "grow an extra nipple"
+      ],
+      1
+    )
   }
 }